[JE2-227] Moved ClientDetails to the components
