+++ frontend/src/features/clients/components/ClientDetails.tsx
